Split up and organize H5 tests
